Move prop test PolicySet to separate file.
